third_party/WebKit/LayoutTests/fast/forms/time-multiple-fields/time-multiple-fields-stepup-stepdown-from-renderer-hour-expected.txt

Check stepping-up and -down for time input fields from renderer. No cases of empty initial values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Hours, 1-12
PASS stepUp("07:00", 1, null, null) is "08:00"
PASS stepDown("07:00", 1, null, null) is "06:00"
PASS stepUp("11:00", 1, null, null) is "00:00"
PASS stepDown("00:00", 1, null, null) is "11:00"
PASS stepUp("06:00", 7200, null, null) is "08:00"
PASS stepDown("06:00", 7200, null, null) is "04:00"
PASS stepUp("10:00", 7200, null, null) is "00:00"
PASS stepDown("00:00", 7200, null, null) is "10:00"
PASS stepUp("07:00", 7200, null, null) is "08:00"
PASS stepDown("07:00", 7200, null, null) is "06:00"
PASS stepUp("06:00", 3601, null, null) is "07:00"
PASS stepDown("06:00", 3601, null, null) is "05:00"
PASS stepUp("06:00", 0, null, null) is "07:00"
PASS stepDown("06:00", 0, null, null) is "05:00"
PASS stepUp("06:00", 86400, null, null) is "00:00"
PASS stepDown("06:00", 86400, null, null) is "00:00"
PASS stepUp("06:00", 36000, null, null) is "07:00"
PASS stepDown("06:00", 36000, null, null) is "05:00"
PASS stepUp("06:00", 7200, "01:00", null) is "07:00"
PASS stepDown("06:00", 7200, "01:00", null) is "05:00"
PASS test("06:00", 7200, null, null, ["Delete", "ArrowUp"]) is "02:00"
PASS test("06:00", 7200, null, null, ["Delete", "ArrowDown"]) is "00:00"
PASS getUserAgentShadowTextContent(input) is "12:00 AM"
PASS test("06:00", 7200, "01:00", null, ["Delete", "ArrowUp"]) is "01:00"
PASS test("06:00", 7200, "01:00", null, ["Delete", "ArrowDown"]) is "11:00"
PASS stepUp("17:00", 1, "17:00", "20:00") is "18:00"
PASS stepDown("17:00", 1, "17:00", "20:00") is "20:00"
PASS stepUp("17:00", 1, "15:00", "17:00") is "15:00"
PASS stepDown("17:00", 1, "15:00", "17:00") is "16:00"
PASS stepUp("15:00", 1, "17:00", "20:00") is "17:00"
PASS stepDown("15:00", 1, "17:00", "20:00") is "20:00"
PASS stepUp("15:00", 1, "13:00", "13:00") is "13:00"
PASS stepDown("15:00", 1, "13:00", "13:00") is "13:00"
PASS stepUp("12:00", 1, "12:00", "15:00") is "13:00"
PASS stepDown("12:00", 1, "12:00", "15:00") is "23:00"
PASS stepUp("15:00", 1, "12:00", "15:00") is "16:00"
PASS stepDown("15:00", 1, "12:00", "15:00") is "14:00"
PASS stepUp("12:00", 1, "10:00", "12:00") is "13:00"
PASS stepDown("12:00", 1, "10:00", "12:00") is "23:00"
PASS stepUp("00:00", 1, "00:00", "03:00") is "01:00"
PASS stepDown("00:00", 1, "00:00", "03:00") is "11:00"
PASS stepUp("15:00", 1, "10:00", "15:00") is "16:00"
PASS stepDown("10:00", 1, "10:00", "15:00") is "09:00"
PASS stepUp("17:00", 7200, "17:00", "20:00") is "19:00"
PASS stepDown("17:00", 7200, "17:00", "20:00") is "19:00"
PASS stepUp("17:00", 7200, "17:00", "18:00") is "17:00"
PASS stepDown("17:00", 7200, "17:00", "18:00") is "17:00"
Hours, 0-11
PASS stepUp("11:00", 1, null, null) is "00:00"
PASS getUserAgentShadowTextContent(input) is "00:00 AM"
PASS stepDown("00:00", 1, null, null) is "11:00"
PASS getUserAgentShadowTextContent(input) is "11:00 AM"
PASS stepUp("23:00", 1, null, null) is "12:00"
PASS getUserAgentShadowTextContent(input) is "00:00 PM"
PASS stepDown("12:00", 1, null, null) is "23:00"
PASS getUserAgentShadowTextContent(input) is "11:00 PM"
PASS test("06:00", 7200, null, null, ["Delete", "ArrowUp"]) is "00:00"
PASS test("06:00", 7200, null, null, ["Delete", "ArrowDown"]) is "10:00"
PASS test("06:00", 7200, "01:00", null, ["Delete", "ArrowUp"]) is "01:00"
PASS test("06:00", 7200, "01:00", null, ["Delete", "ArrowDown"]) is "11:00"
PASS stepUp("17:00", 1, "17:00", "20:00") is "18:00"
PASS stepDown("17:00", 1, "17:00", "20:00") is "20:00"
PASS stepUp("17:00", 1, "15:00", "17:00") is "15:00"
PASS stepDown("17:00", 1, "15:00", "17:00") is "16:00"
PASS stepUp("15:00", 1, "17:00", "20:00") is "17:00"
PASS stepDown("15:00", 1, "17:00", "20:00") is "20:00"
PASS stepUp("15:00", 1, "13:00", "13:00") is "13:00"
PASS stepDown("15:00", 1, "13:00", "13:00") is "13:00"
PASS stepUp("12:00", 1, "12:00", "15:00") is "13:00"
PASS stepDown("12:00", 1, "12:00", "15:00") is "15:00"
PASS stepUp("15:00", 1, "12:00", "15:00") is "12:00"
PASS stepDown("15:00", 1, "12:00", "15:00") is "14:00"
PASS stepUp("12:00", 1, "10:00", "12:00") is "13:00"
PASS stepDown("12:00", 1, "10:00", "12:00") is "23:00"
PASS stepUp("00:00", 1, "00:00", "03:00") is "01:00"
PASS stepDown("00:00", 1, "00:00", "03:00") is "03:00"
PASS stepUp("15:00", 1, "10:00", "15:00") is "16:00"
PASS stepDown("10:00", 1, "10:00", "15:00") is "09:00"
PASS stepUp("20:00", 7200, "17:00", "20:00") is "17:00"
PASS stepDown("20:00", 7200, "17:00", "20:00") is "19:00"
Hours, 0-23
PASS stepUp("07:00", 1, null, null) is "08:00"
PASS stepDown("07:00", 1, null, null) is "06:00"
PASS stepUp("23:00", 1, null, null) is "00:00"
PASS stepDown("00:00", 1, null, null) is "23:00"
PASS stepUp("06:00", 7200, null, null) is "08:00"
PASS stepDown("06:00", 7200, null, null) is "04:00"
PASS stepUp("22:00", 7200, null, null) is "00:00"
PASS stepDown("00:00", 7200, null, null) is "22:00"
PASS stepUp("07:00", 7200, null, null) is "08:00"
PASS stepDown("07:00", 7200, null, null) is "06:00"
PASS stepUp("06:00", 3601, null, null) is "07:00"
PASS stepDown("06:00", 3601, null, null) is "05:00"
PASS stepUp("06:00", 0, null, null) is "07:00"
PASS stepDown("06:00", 0, null, null) is "05:00"
PASS stepUp("06:00", 86400, null, null) is "00:00"
PASS stepDown("06:00", 86400, null, null) is "00:00"
PASS stepUp("06:00", 36000, null, null) is "07:00"
PASS stepDown("06:00", 36000, null, null) is "05:00"
PASS stepUp("06:00", 7200, "01:00", null) is "07:00"
PASS stepDown("06:00", 7200, "01:00", null) is "05:00"
PASS test("06:00", 7200, null, null, ["Delete", "ArrowUp"]) is "00:00"
PASS getUserAgentShadowTextContent(input) is "00:00"
PASS test("06:00", 7200, null, null, ["Delete", "ArrowDown"]) is "22:00"
PASS test("06:00", 7200, "01:00", null, ["Delete", "ArrowUp"]) is "01:00"
PASS test("06:00", 7200, "01:00", null, ["Delete", "ArrowDown"]) is "23:00"
PASS stepUp("17:00", 1, "17:00", "20:00") is "18:00"
PASS stepDown("17:00", 1, "17:00", "20:00") is "20:00"
PASS stepUp("17:00", 1, "15:00", "17:00") is "15:00"
PASS stepDown("17:00", 1, "15:00", "17:00") is "16:00"
PASS stepUp("15:00", 1, "17:00", "20:00") is "17:00"
PASS stepDown("15:00", 1, "17:00", "20:00") is "20:00"
PASS stepUp("15:00", 1, "13:00", "13:00") is "13:00"
PASS stepDown("15:00", 1, "13:00", "13:00") is "13:00"
PASS stepUp("00:00", 1, "00:00", "03:00") is "01:00"
PASS stepDown("00:00", 1, "00:00", "03:00") is "03:00"
PASS stepUp("03:00", 1, "00:00", "03:00") is "00:00"
PASS stepDown("03:00", 1, "00:00", "03:00") is "02:00"
PASS stepUp("12:00", 1, "10:00", "12:00") is "10:00"
PASS stepDown("12:00", 1, "10:00", "12:00") is "11:00"
PASS stepUp("00:00", 1, "00:00", "03:00") is "01:00"
PASS stepDown("00:00", 1, "00:00", "03:00") is "03:00"
PASS stepUp("15:00", 1, "10:00", "15:00") is "10:00"
PASS stepDown("10:00", 1, "10:00", "15:00") is "15:00"
PASS stepUp("20:00", 7200, "17:00", "20:00") is "17:00"
PASS stepDown("20:00", 7200, "17:00", "20:00") is "19:00"
Hours, 1-24
PASS stepUp("11:00", 1, null, null) is "12:00"
PASS getUserAgentShadowTextContent(input) is "12:00"
PASS stepDown("00:00", 1, null, null) is "23:00"
PASS getUserAgentShadowTextContent(input) is "23:00"
PASS stepUp("23:00", 1, null, null) is "00:00"
PASS getUserAgentShadowTextContent(input) is "24:00"
PASS stepDown("12:00", 1, null, null) is "11:00"
PASS getUserAgentShadowTextContent(input) is "11:00"
PASS test("06:00", 7200, null, null, ["Delete", "ArrowUp"]) is "02:00"
PASS test("06:00", 7200, null, null, ["Delete", "ArrowDown"]) is "00:00"
PASS test("06:00", 7200, "01:00", null, ["Delete", "ArrowUp"]) is "01:00"
PASS test("06:00", 7200, "01:00", null, ["Delete", "ArrowDown"]) is "23:00"
PASS stepUp("17:00", 1, "17:00", "20:00") is "18:00"
PASS stepDown("17:00", 1, "17:00", "20:00") is "20:00"
PASS stepUp("17:00", 1, "15:00", "17:00") is "15:00"
PASS stepDown("17:00", 1, "15:00", "17:00") is "16:00"
PASS stepUp("15:00", 1, "17:00", "20:00") is "17:00"
PASS stepDown("15:00", 1, "17:00", "20:00") is "20:00"
PASS stepUp("15:00", 1, "13:00", "13:00") is "13:00"
PASS stepDown("15:00", 1, "13:00", "13:00") is "13:00"
PASS stepUp("00:00", 1, "00:00", "03:00") is "01:00"
PASS stepDown("00:00", 1, "00:00", "03:00") is "23:00"
PASS stepUp("03:00", 1, "00:00", "03:00") is "04:00"
PASS stepDown("03:00", 1, "00:00", "03:00") is "02:00"
PASS stepUp("12:00", 1, "10:00", "12:00") is "10:00"
PASS stepDown("12:00", 1, "10:00", "12:00") is "11:00"
PASS stepUp("00:00", 1, "00:00", "03:00") is "01:00"
PASS stepDown("00:00", 1, "00:00", "03:00") is "23:00"
PASS stepUp("15:00", 1, "10:00", "15:00") is "10:00"
PASS stepDown("10:00", 1, "10:00", "15:00") is "15:00"
PASS stepUp("17:00", 7200, "17:00", "20:00") is "19:00"
PASS stepDown("17:00", 7200, "17:00", "20:00") is "19:00"
PASS stepUp("17:00", 7200, "17:00", "18:00") is "17:00"
PASS stepDown("17:00", 7200, "17:00", "18:00") is "17:00"

PASS successfullyParsed is true

TEST COMPLETE
